refactor: Moved API constanst to dblp_api.py
